feat: effect composer sfc
