+++ playground/.eslintrc-auto-import.json
@@ -58,7 +58,6 @@
     "watch": true,
     "watchEffect": true,
     "watchPostEffect": true,
-    "watchSyncEffect": true,
-    "toValue": true
+    "watchSyncEffect": true
   }
 }
