+++ src/index.ts
@@ -1,4 +1,4 @@
-import { EffectComposer } from './core/useEffectComposer/component'
+import EffectComposer from './core/EffectComposer.vue'
 import { Bloom } from './core/effects/Bloom'
 import { Glitch } from './core/effects/Glitch'
 import { Outline } from './core/effects/Outline'
